At a functional stage
`dtreetrawl` trawls/traverses directory tree or file path to collect
stats of every entry in the tree.

Stats of all paths passed as arguments to `dtreetrawl` is printed to
the standard output, errors to standard error. After the stats of
trawled entries, a summary of that run for the root path follows. When
multiple paths are given as arguments, stats continue to print to the
standard output.

In the terse output, each `trawlentry`/`tentry`/`tent` is surrounded by
the set delimiter/separator twice. That is, say if the delimiter is
`:`, then, at the begining and end of the tentry, `::` will be placed;
fields within the tentry will be separated by `:`. A successful run is
followed by a summary(`dstat`), to differentiate it from the tentries,
it will have only a single delimiter at the begining and ending.
`trawlentry` starts with a double delimiter, `dstat` starts with a
single delimiter.

`tentry` fields:

 - entry/path name
 - base name of the path
 - level from root
 - file type
 - referent name if it's a symlink
 - file size in bytes
 - inode number
 - number of directory entries
 - permission in octal notation
 - hard link count
 - ownership: user and group id
 - I/O block size
 - allocated blocks
 - last status change (ctime)
 - last file access (atime)
 - last file modification (mtime)
 - hash

`dstat` fields:

 - root path name of this run
 - elapsed time in seconds
 - start time
 - root hash
 - hash type
 - depth
 - total size
 - total entries
 - total directories
 - total regular files
 - total symlinks
 - total block devices
 - total char devices
 - total sockets
 - total FIFOs/pipes

For the terse output:

Fields appear in the same order as listed. Fields that are not
applicable are left blank, which leads to adjacent delimiters.
After each run, a new line is printed after dstat(summary
 entry) to help differentiate the next run for another root path.
Units are not attached to the field values in terse output.
Time values are in UTC as epoch for terse output(local string form
for human readable output) Sizes are all in bytes.

File types:
SOCK  socket
LNK   symbolic link
REG   regular file
BLK   block device
DIR   directory
CHR   character device
FIFO  FIFO/pipes

Signed-off-by: six-k <ramsri.hp@gmail.com>